+++ wp-content/plugins/themer-pro/lib/admin/update/update.php
@@ -127,6 +127,14 @@ function themer_pro_update() {
 	// Update to Themer Pro 1.4.2
 	if ( version_compare( get_option( 'themer_pro_version_number' ), '1.4.2', '<' ) )
 		update_option( 'themer_pro_version_number', '1.4.2' );
+
+	// Update to Themer Pro 1.5.0
+	if ( version_compare( get_option( 'themer_pro_version_number' ), '1.5.0', '<' ) )
+		update_option( 'themer_pro_version_number', '1.5.0' );
+
+	// Update to Themer Pro 1.5.1
+	if ( version_compare( get_option( 'themer_pro_version_number' ), '1.5.1', '<' ) )
+		update_option( 'themer_pro_version_number', '1.5.1' );
 	
 	// Finalize the update sequence.
 	themer_pro_activate_post();